% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{avg_inds}
\alias{avg_inds}
\title{Averaged intersection between multiple trajectories and segmented line}
\usage{
avg_inds(SpLlst)
}
\arguments{
\item{SpLlst}{A list of animal trajectory of class ltraj containing multiple individuals}
}
\value{
A  SpatialLinesDataFrame object
}
\description{
Combined a list list of corriIntersects object to obtain population level averaged statistics regarding crossing count
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
 t3<-corriIntersects_All(albatross, t2) 
t4<-avg_inds(t3) 
}
\keyword{SegmentSpL}
\keyword{ltraj}
\keyword{plotcorri_ind}
