% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{corriIntersects}
\alias{corriIntersects}
\title{Intersection between individual trajectory and segmented line}
\usage{
corriIntersects(traj, corri, per_day = TRUE, plot = F)
}
\arguments{
\item{traj}{An animal trajectory of class ltraj}

\item{corri}{A segmented SpatialLines as returned bt function SegmentSpL}

\item{per_day}{Specify if the number of crossing should be standardized over the time (in days) the animal was monitored or over the total number of locations, default=TRUE. Long gap in data should be identified as separate burst}

\item{plot}{Whether a plot should be returned (default=F)}
}
\value{
A SpatialLinesDataFrame object
}
\description{
Calculate summary statistics regarding frequency of intersections between an animal trajectory and a segmented SpatialLines object
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
t3<-corriIntersects(albatross[3], t2, plot=T)
 plot(ltraj2sldf(albatross[3]), add=T)
}
\keyword{SegmentSpL}
\keyword{ltraj}
\keyword{plotcorri_ind}
