% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{corriIntersects_All}
\alias{corriIntersects_All}
\title{Intersection between multiple trajectories and segmented line}
\usage{
corriIntersects_All(trajs, corri)
}
\arguments{
\item{corri}{A segmented SpatialLines as returned bt function SegmentSpL}

\item{traj}{An animal trajectory of class ltraj containing multiple individuals}
}
\value{
A list of SpatialLinesDataFrame object
}
\description{
Wrapper function that apply corriIntersects to all individual in a trajectory object
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
t3<-corriIntersects_All(albatross, t2) 
}
\keyword{SegmentSpL}
\keyword{ltraj}
\keyword{plotcorri_ind}
