% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{hr_split}
\alias{hr_split}
\title{Intersection of a home-range (polygon) with linear features}
\usage{
hr_split(pol = hr, line = corri)
}
\arguments{
\item{pol}{A SpatialPolygons* object representing an animal home-range.}

\item{line}{A linear feature of class SpatialLines}
}
\value{
A list containing the segmented polygon, the % area of the bigger segment, the number of segment, and polygon initial total area
}
\description{
Intersects a SpatialPolygons* object with a SpatialLines* object and return a divided polygon (when overlap) with summary metric regarding % of area left (bigger sectio), number of segments created, and total initial area of the polygon. Taken from: http://r-sig-geo.2731867.n2.nabble.com/Split-polygon-by-line-td7588625.html
}
\examples{
pol<-SpatialPolygons(list(Polygons(list(Polygon(cbind(c(0,1,1,0,0),c(0,0,1,1,0)))), ID="polygon"))) 
line<-SpatialLines(list(Lines(list(Line(cbind(c(0,1),c(0.4,0.4)))), ID="line"))) 
splt<-hr_split(pol, line)
plot(splt[[1]])
splt[2:4] 
}
\keyword{SpatialLines}
\keyword{SpatialPolygons}
