% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{match_pts}
\alias{match_pts}
\title{Associate closest elements between two SpatialPoints features}
\usage{
match_pts(pts1, pts2)
}
\arguments{
\item{pts1}{A SpatialLines* or SpatialPoints object. If a SpatialLines is provided, the object is first converted to point using getSpatialLinesMidPoints}

\item{pts2}{A SpatialPoints* object}
}
\description{
Evaluate the distance between a SpatialPoints/Lines* object and another SpatialPoints* object and return which features in first object that are the closest to each feature in second object
}
\examples{
 x <- c(1,5,4,8)
y <- c(1,3,4,7)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=10, merge.last=F)
 Pts<-SpatialPoints(matrix(c(1,3,7,1,2,7), nrow=3, ncol=2))
 match_pts(t2, Pts)
}
\keyword{SpatialPoints}
\keyword{match}
