% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{optim_corri}
\alias{optim_corri}
\title{Optimization of wildlife crossing locations over a linear features}
\usage{
optim_corri(corri, var = 4, n = 5, pct_keep = 1, nb_ind = 2, ln = F,
  rm = NULL, add = NULL, weight = 0.5, plot = T, time_limit = -1,
  gap_limit = -1, first_feasible = F, ...)
}
\arguments{
\item{corri}{A segmented SpatialLines object returned by avg_inds}

\item{var}{Variable used to represent importance of segment for animal. Number refer to columns of dataframe produced by avg_inds (default = 4, average percent crossing)}

\item{n}{Number of crossing to place (default =5) in addition to fixed points (i.e. if a SpatialPoints* object is provided to the add argument)}

\item{pct_keep}{Percentage of segment to consider in optimization. Removal is based on percentile of values of importance of segment for animal crossing}

\item{ln}{Whether the natural logarithm of the variable value should be taken. Default=F}

\item{rm}{A SpatialPoints object indicating the location that most be excluded from the optimization (defaul is NULL)}

\item{add}{A SpatialPoints object indicating the location where a crossing is already present, or must be place to this location. Number of points included here will be added to n to give the total number of crossing selected.}

\item{weight}{Argument setting the weight given to the spatial spread relative to the crossing importance (default = 0.5 meaning equal importance)}

\item{plot}{Whether a plot showing the crossing location should be returned (default=T)}

\item{...}{additional arguments that can be specify to Rsymphony_solve_lp (time_limit, gap_limit, first_feasible)}

\item{cost}{A vector of same length that the number of segment in corri giving the cost for each segment}
}
\value{
A list containing the segmented polygon, the % area of the bigger segment, the number of segment, and polygon initial total area
}
\description{
The function use linear programming to optimize the location of wildlife crossing over a linear feature. The function maximise the spatial spread of locations, 
   and the importance of specific location for animal crossing. The user can specify the number of crossing location desired, if some segment should be excluded, or if the location
    of some crossings are already decided. The user also need to specify the weight given to the spatial argument and the importance of crossing (default is equal importance to each).
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
 t3<-corriIntersects_All(albatross, t2) 
t4<-avg_inds(t3)
 opti1<-optim_corri(t4, var=4, n=3, nb_ind=1, weight=0.5, plot=T) #Equal weight
 opti2<-optim_corri(t4, var=4, n=3, nb_ind=1, weight=0.95, plot=T) #More importance to spatial spread
}
\keyword{SpatialLines}
\keyword{SpatialPolygons}
