% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{optim_mclp}
\alias{optim_mclp}
\title{Optimization of wildlife crossing locations over a linear features using the maximum coverage location problem}
\usage{
optim_mclp(corri_ls, n = 5, dist = 10 * 1000, rm = NULL, add = NULL,
  plot = T, time_limit = -1, gap_limit = -1, first_feasible = F, ...)
}
\arguments{
\item{corri_ls}{A segmented SpatialLines list returned by corriIntersects_All}

\item{n}{Number of crossing to place (default =5) in addition to fixed points (i.e. if a SpatialPoints* object is provided to the add argument)}

\item{dist}{Distance used to considered a given segment as covered (ie radius)}

\item{rm}{A SpatialPoints object indicating the location that most be excluded from the optimization (defaul is NULL)}

\item{add}{A SpatialPoints object indicating the location where a crossing is already present, or must be place to this location. Number of points included here will be added to n to give the total number of crossing selected.}

\item{plot}{Whether a plot showing the crossing location should be returned (default=T)}

\item{...}{additional arguments that can be specify to Rsymphony_solve_lp (time_limit, gap_limit, first_feasible)}
}
\value{
A list containing the segmented polygon, the % area of the bigger segment, the number of segment, and polygon initial total area
}
\description{
The function use linear programming to optimize the location of wildlife crossing over a linear feature. The function maximise the number of individuals the selected features will assist.
 The user can specify the number of crossing location desired, a coverage distance,  if some segment should be excluded, or if the location
    of some crossings are already decided.
}
\examples{
 require(adehabitatLT)
 x <- c(0,0)
 y <- c(-6500000,-4500000)
 t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
 t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
 data (albatross) #From package adehabitatLT
 t3<-corriIntersects_All(albatross, t2) 
 test<-optim_mclp(t3, n=2,dist=10*1000, plot=T)
}
\keyword{SpatialLines}
\keyword{SpatialPolygons}
