% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plot_optim}
\alias{plot_optim}
\title{Plot of selected crossing structures following optimization using ILP}
\usage{
plot_optim(corri, var = 4, optim, main = "Default", ...)
}
\arguments{
\item{corri}{A SpatialLinesDataFrame object returned by avg_inds}

\item{var}{Variable to display (default =4) 1= Average count, 2=Total count, 3=Average count standardized by length of tracking, 4=Average count standardized by length of tracking and spatial sampling bias, 5= Number of different individuals}

\item{optim}{Output of the optimization function optim_corri}
}
\value{
A plot
}
\description{
Produce a color-coded plot of density of crossing of multiple individualsand optimal crossing structure. Used to display result of function optim_corri
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
t3<-corriIntersects_All(albatross, t2)
t4<-avg_inds(t3) 
 opti1<-optim_corri(t4, var=4, n=3, nb_ind=1, weight=0.5, plot=T) #Equal weight 
 plot_optim(t4,  var=4, opti1, main="Crossings") 
}
\keyword{avg_inds}
\keyword{corriIntersects_All}
\keyword{optim_corri}
