% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plotcorri_grp}
\alias{plotcorri_grp}
\title{Plot of density of crossing of multiple individuals with linear features}
\usage{
plotcorri_grp(SpL, nb_breaks = 5, var = 4, main = "Default")
}
\arguments{
\item{SpL}{A SpatialLinesDataFrame object returned by avg_inds}

\item{nb_breaks}{The number of breaks to use in display (default =5), must be <=10}

\item{var}{Variable to display (default =4) 1= Average count, 2=Total count, 3=Average count standardized by length of tracking, 4=Average count standardized by length of tracking and spatial sampling bias, 5= Number of different individuals}
}
\value{
A plot
}
\description{
Produce a color-coded plot of density of crossing of multiple individuals with a segmented linear features. Used to display result of function avg_inds
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
t3<-corriIntersects_All(albatross, t2)
t4<-avg_inds(t3)  
 plotcorri_grp(t4, nb_breaks=5, var=4, main="Albatross")
}
\keyword{avg_inds}
\keyword{corriIntersects_All}
