% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{plotcorri_ind}
\alias{plotcorri_ind}
\title{Plot of density of crossing of an individual with linear features}
\usage{
plotcorri_ind(SpL, nb_breaks = 5, extent = NULL)
}
\arguments{
\item{SpL}{A SpatialLinesDataFrame object returned by corriIntersects}

\item{nb_breaks}{The number of breaks to use in display (default =5), must be <=10}

\item{extent}{If not NULL, an extent object (from package raster) that specify boundaries of the display.}
}
\value{
A plot
}
\description{
Produce a color-coded plot of density of crossing of an individual with a segmented linear features. Used to display result of function corriIntersects
}
\examples{
require(adehabitatLT)
 x <- c(0,0)
y <- c(-6500000,-4500000)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=20, merge.last=F)
data (albatross) #From package adehabitatLT
t3<-corriIntersects(albatross[3], t2, plot=F)
plotcorri_ind(t3, nb_breaks=4, extent=raster::extent(ltraj2spdf(albatross)))
}
\keyword{corriIntersects}
