% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts.R
\name{SegmentSpL}
\alias{SegmentSpL}
\title{Segmentation of SpatialLines object into muliple segments}
\usage{
SegmentSpL(sl, length = 0, n.parts = 0, merge.last = FALSE)
}
\arguments{
\item{sl}{SpatialLines object}

\item{length}{Length of individual segment (in units of sl)}

\item{n.parts}{Alternatively, the number of segments to create}

\item{merge.last}{Whether the last segment (of a different length) should be merged with the previous segment}
}
\value{
A SpatialLines object
}
\description{
Segment a SpatialLines object in segment of equal length (taken from: http://rstudio-pubs-static.s3.amazonaws.com/10685_1f7266d60db7432486517a111c76ac8b.html)
}
\examples{
 x <- c(1,5,4,8)
y <- c(1,3,4,7)
t1<-SpatialLines(list(Lines(Line(cbind(x,y)), ID="a")))
t2<-SegmentSpL(t1, n.parts=10, merge.last=F)
	plot(t2, col = rep(c(1, 2), length.out = length(t2)), axes = T)
}
\keyword{CreateSegment}
\keyword{CreateSegments}
\keyword{MergeLast}
\keyword{SegmentSpL}
